backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
alembic
psycopg2-binary
sqlalchemy>=2.0
asyncpg
greenlet
pydantic-settings
email-validator
